requirements.txt
pdfminer.six
lxml
mysql-connector-python
requests